[Silver III] Title: 주유소, Time: 92 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -12,14 +12,14 @@ int main() {
 	int N;
 	cin >> N;
 
-	vector<int>road(N - 1);
+	vector<long>road(N - 1);
 	for (int i = 0; i < N - 1; i++) cin >> road[i];
 
-	vector<int>oil(N);
+	vector<long>oil(N);
 	for (int i = 0; i < N; i++) cin >> oil[i];
 
-	int result = 0;
-	int o = oil[0];//처음은 무조건 오일 사용해야함
+	long result = 0;
+	long o = oil[0];//처음은 무조건 오일 사용해야함
 
 	for (int i = 0; i < N - 1; i++) {
 		o = min(o, oil[i]);
